Unified trace format into a single header
